requirements.txt
influx_line_protocol~=0.1.4
Telethon~=1.24.0
PyYAML~=6.0